requirements.txt
qdrant-client
sentence-transformers
numpy
geopy
folium
streamlit-folium
nltk
scikit-learn
joblib
torch